Create RVersion
I thought R would be special because it seems like it's a data scientist focused language. That's not how languages work I reckon. Too magical. This looks like it would also take hundreds of thousands of years to finish and I ain't got time for that mister. At least it makes a chart when interrupted... albeit not one I was looking for, but I didn't realize that until typing this now... *sigh*. You want something done right, you better not do it yourself. Help. Thinking of possible solutions... hmmmm